lambda/UploadImage/upload-image/requirements.txt
requests
requests_aws4auth
opensearch-py
pillow